@@ -7,34 +7,36 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "deploy": "yarn lint && yarn format:check && yarn build",
     "check-types": "tsc --noemit",
-    "lint:ts": "eslint . --ext .ts,.tsx --fix",
-    "lint": "yarn check-types && yarn lint:ts"
+    "lint:ts": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "lint": "yarn check-types && yarn lint:ts",
+    "format:check": "prettier . --check",
+    "format:fix": "prettier . --write"
   },
   "devDependencies": {
-    "@types/dom-webcodecs": "^0.1.11",
-    "@types/node": "^20.14.11",
-    "@types/react": "^18.3.3",
+    "@types/node": "^22.5.5",
+    "@types/react": "^18.3.8",
     "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.16.1",
-    "@typescript-eslint/parser": "^7.16.1",
     "@vitejs/plugin-react": "^4.3.1",
-    "@webgpu/types": "^0.1.44",
-    "autoprefixer": "^10.4.19",
-    "eslint": "8.57.0",
+    "@webgpu/types": "^0.1.46",
+    "autoprefixer": "^10.4.20",
+    "eslint": "9.11.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-react": "^7.34.4",
+    "eslint-plugin-react": "^7.36.1",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "postcss": "^8.4.39",
+    "postcss": "^8.4.47",
     "prettier": "^3.3.3",
-    "tailwindcss": "^3.4.6",
-    "typescript": "^5.5.3",
-    "vite": "^5.3.4"
+    "tailwindcss": "^3.4.12",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.6.0",
+    "vite": "^5.4.7"
   },
   "dependencies": {
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "wgpu-matrix": "^3.0.1"
+    "wgpu-matrix": "^3.2.0"
   }
 }
