@@ -1,6 +1,18 @@
 module.exports = {
-  arrowParens: 'avoid',
-  printWidth: 140,
+  arrowParens: 'always',
+  bracketSameLine: true,
+  bracketSpacing: true,
+  endOfLine: 'lf',
+  htmlWhitespaceSensitivity: 'css',
+  insertPragma: false,
+  jsxSingleQuote: false,
+  printWidth: 120,
+  proseWrap: 'preserve',
+  quoteProps: 'as-needed',
+  requirePragma: false,
+  semi: true,
   singleQuote: true,
+  tabWidth: 2,
   trailingComma: 'all',
+  useTabs: false,
 };
