@@ -4,7 +4,11 @@ const NoWebGPU = () => {
   return (
     <div className="text-red-600 text-xl font-bold p-5">
       Your browser does not support WebGPU yet{' '}
-      <a className="underline" target="_blank" rel="noreferrer" href="https://github.com/gpuweb/gpuweb/wiki/Implementation-Status">
+      <a
+        className="underline"
+        target="_blank"
+        rel="noreferrer"
+        href="https://github.com/gpuweb/gpuweb/wiki/Implementation-Status">
         (Implementation Status)
       </a>
       <br />
