@@ -4,5 +4,7 @@ import App from './components/App';
 import './index.css';
 
 const container = document.getElementById('root');
-const root = createRoot(container);
-root.render(<App />);
+if (container) {
+  const root = createRoot(container);
+  root.render(<App />);
+}
