@@ -1,5 +1,6 @@
-// vite.config.ts
 import react from '@vitejs/plugin-react';
+import path from 'node:path';
+import process from 'node:process';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
@@ -8,8 +9,14 @@ export default defineConfig({
     port: 8082,
     host: '0.0.0.0',
   },
+  root: 'src',
+  publicDir: '../public',
   build: {
     minify: true,
+    outDir: '../dist',
+  },
+  resolve: {
+    alias: { '/src': path.resolve(process.cwd(), 'src') },
   },
   base: '',
 });
